backend/requirements.txt
# ─── Core Django ──────────────────────────────────────────────────────────────
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==24.2

# ─── Authentication ───────────────────────────────────────────────────────────
djangorestframework-simplejwt==5.3.1

# ─── WebSockets (real-time board communication) ───────────────────────────────
channels==4.1.0
channels-redis==4.2.0
daphne>=4.1.2        # ASGI server that runs Django Channels

# ─── Task Queue (ad scheduler) ────────────────────────────────────────────────
celery==5.4.0
redis==5.0.4           # Python Redis client (used by both Channels and Celery)

# ─── Database ─────────────────────────────────────────────────────────────────
psycopg2-binary>=2.9.9  # PostgreSQL adapter

# ─── File Handling ────────────────────────────────────────────────────────────
Pillow>=10.3.0          # Image processing (thumbnails, validation)
# ─── Payments ─────────────────────────────────────────────────────────────────
stripe>=8.0.0

python-dotenv>=1.0.0
django-storages==1.14.2
boto3==1.34.69


